refactor(components): [row] switch to script-setup syntax (#7961)
* refactor(components):[row] switch to script-setup syntax

* refactor(components): [row] export const type
