+++ packages/components/row/index.ts
@@ -1,5 +1,5 @@
 import { withInstall } from '@element-plus/utils'
-import Row from './src/row'
+import Row from './src/row.vue'
 
 export const ElRow = withInstall(Row)
 export default ElRow
